Bugfix: SeqDir run overflow
u8 has a max value of 255.
Many sequencing runs go into the thousands, so there will be overflows when trying to process later runs on a machine.
Increasing the size of a run from u8 to u16 should increase the max value to 65535 and be more than sufficient.

+++ src/data/mod.rs
@@ -28,7 +28,7 @@ struct SeqDir {
     path: PathBuf,
     date: Date,
     instrument: String,
-    run: u8,
+    run: u16,
     position: char,
     flowcell: String,
     description: String,
@@ -69,7 +69,7 @@ impl SeqDir {
                 )
                 .unwrap();
                 let instrument = cap.get(4).unwrap().as_str();
-                let run = cap.get(5).unwrap().as_str().parse::<u8>().unwrap();
+                let run = cap.get(5).unwrap().as_str().parse::<u16>().unwrap();
                 let pos = cap.get(6).unwrap().as_str().parse::<char>().unwrap();
                 let flowcell = cap.get(7).unwrap().as_str();
                 let description = cap.get(8).unwrap().as_str();
@@ -113,7 +113,7 @@ impl SeqDir {
     }
 
     /// Run number of SeqDir run
-    pub fn run(&self) -> u8 {
+    pub fn run(&self) -> u16 {
         self.run
     }
